Rafadana/event price (#52)
* feat: :sparkles: show event fee in event content card and swipcard

* feat: :sparkles: Add price on event detail

* feat: :sparkles: show price on eventList component

* refactor: :lipstick: update price design on EventList component

* feat: :sparkles: show event fee in event content card and swipcard

* feat: :sparkles: Add price on event detail

* feat: :sparkles: show price on eventList component

* refactor: :lipstick: update price design on EventList component

* style: :lipstick: add commas

+++ lib/event/data/common/event_price_request_model.dart
@@ -5,7 +5,7 @@ part 'event_price_request_model.g.dart';
 
 @JsonSerializable()
 class EventPriceRequestModel extends BaseModel {
-  final int fee;
+  final double fee;
   final String currency;
   const EventPriceRequestModel({required this.fee, required this.currency});
